build: DH-20465: Fixes for MethodListValidator dependencies. (#7222)
