+++ server/build.gradle
@@ -79,7 +79,7 @@ dependencies {
     }
 
     implementation project(':extensions-validation-base')
-    implementation project(':extensions-validation-method-list')
+    implementation project(path: ':extensions-validation-method-list', configuration: 'shadow')
 
     api(project(':plugin')) {
         because 'downstream dagger compile, see deephaven-core#1722'
@@ -111,7 +111,7 @@ dependencies {
     testRuntimeOnly project(':log-to-slf4j')
     testRuntimeOnly libs.slf4j.simple
 
-    runtimeOnly project(':authentication:example-providers:psk')
+    runtimeOnly project(path: ':authentication:example-providers:psk', configuration: 'shadow')
 }
 
 TestTools.addEngineOutOfBandTest(project)
