+++ server/src/test/java/io/deephaven/server/table/TestColumnExpressionValidator.java
@@ -34,8 +34,6 @@ import org.apache.commons.lang3.mutable.MutableObject;
 import org.junit.Assert;
 import org.junit.Rule;
 import org.junit.Test;
-import org.openrewrite.java.MethodMatcher;
-import org.openrewrite.java.tree.JavaType;
 
 import java.io.IOException;
 import java.lang.reflect.Constructor;
@@ -1012,66 +1010,6 @@ public class TestColumnExpressionValidator {
         Assert.assertTrue(validator2.permitConstructor(BigInteger.class.getConstructor(String.class)));
         Assert.assertNull(validator2.permitConstructor(BigInteger.class.getConstructor(byte[].class)));
         Assert.assertNull(validator2.permitConstructor(BigDecimal.class.getConstructor(String.class)));
-    }
-
-    @Test
-    public void testMethodCachingBehavior() throws NoSuchMethodException {
-        final List<String> disallowedMethods = List.of("java.lang.StringBuilder toString()");
-        final List<String> allowedMethods = List.of("java.lang.Object toString()");
-
-        final MutableInt allowedMethodCount = new MutableInt(0);
-        final MutableInt disallowedMethodCount = new MutableInt(0);
-        final MutableInt nullMethodCount = new MutableInt(0);
-
-        final MethodListInvocationValidator validator = new MethodListInvocationValidator(allowedMethods) {
-            final List<MethodMatcher> disallowedMatchers =
-                    disallowedMethods.stream().map(MethodMatcher::new).collect(Collectors.toUnmodifiableList());
-
-            @Override
-            public Boolean permitMethod(Method method) {
-                final JavaType.Method jtm = toJavaType(method);
-
-                if (disallowedMatchers.stream().anyMatch(mm -> mm.matches(jtm))) {
-                    disallowedMethodCount.increment();
-                    return false;
-                }
-                Boolean result = super.permitMethod(method);
-                if (result == null) {
-                    nullMethodCount.increment();
-                } else if (!result) {
-                    throw new IllegalStateException();
-                } else {
-                    allowedMethodCount.increment();
-                }
-                return result;
-            }
-        };
-        final CachingMethodInvocationValidator cachingValidator = new CachingMethodInvocationValidator(validator);
-
-        final Method integerToString = Integer.class.getMethod("toString");
-        final Method toUnsignedLong = Integer.class.getMethod("toUnsignedLong", int.class);
-        final Method stringBuilderToString = StringBuilder.class.getMethod("toString");
-
-        Assert.assertTrue(cachingValidator.permitMethod(integerToString));
-        Assert.assertEquals(1, allowedMethodCount.get());
-        Assert.assertEquals(0, disallowedMethodCount.get());
-        Assert.assertEquals(0, nullMethodCount.get());
-        Assert.assertFalse(cachingValidator.permitMethod(stringBuilderToString));
-        Assert.assertEquals(1, allowedMethodCount.get());
-        Assert.assertEquals(1, disallowedMethodCount.get());
-        Assert.assertEquals(0, nullMethodCount.get());
-        Assert.assertNull(cachingValidator.permitMethod(toUnsignedLong));
-        Assert.assertEquals(1, allowedMethodCount.get());
-        Assert.assertEquals(1, disallowedMethodCount.get());
-        Assert.assertEquals(1, nullMethodCount.get());
-
-        // verify we did not need to ask the validator a second time
-        Assert.assertTrue(cachingValidator.permitMethod(integerToString));
-        Assert.assertFalse(cachingValidator.permitMethod(stringBuilderToString));
-        Assert.assertNull(cachingValidator.permitMethod(toUnsignedLong));
-        Assert.assertEquals(1, allowedMethodCount.get());
-        Assert.assertEquals(1, disallowedMethodCount.get());
-        Assert.assertEquals(1, nullMethodCount.get());
     }
 
     @Test
